requirements.txt
# Core database dependencies
psycopg==3.2.6
SQLAlchemy==2.0.36

# Environment and configuration
python-dotenv==1.0.1

# Logging and utilities
rich==13.9.4

# Testing (optional - for development)
pytest==8.3.4
pytest-flask==1.3.0